configure eslint - report coding-style violations [Finishes #171923628]
