--- .eslintrc.js
@@ -1,30 +1,0 @@
-module.exports = {
-  env: {
-    es6: true,
-    node: true
-  },
-  extends: ['airbnb-base'],
-  globals: {
-    Atomics: 'readonly',
-    SharedArrayBuffer: 'readonly'
-  },
-  parserOptions: {
-    ecmaVersion: 2018,
-    sourceType: 'module'
-  },
-  rules: {
-    indent: ['error', 2],
-    'comma-dangle': [
-      'error',
-      {
-        arrays: 'never',
-        objects: 'never',
-        imports: 'never',
-        exports: 'never',
-        functions: 'never'
-      }
-    ],
-    'operator-linebreak': [2, 'after', { overrides: { '?': 'after' } }],
-    'function-paren-newline': 'off'
-  }
-};
