+++ .hound.yml
@@ -1,0 +1,3 @@
+eslint:
+  enabled: true
+  config_file: .eslintrc.json
