+++ src/models/lost.js
@@ -1,6 +1,28 @@
-import mongoose from 'mongoose';
+import mongoose, { Schema } from 'mongoose';
+import config from '../../config/config';
 
-const Schema = mongoose.Schema;
+const {
+  db: { host, port, name }
+} = config;
+export const url =
+  `mongodb://${host}:${port}/${name}?authSource=admin` ||
+  config.db.database_url;
+
+const options = {
+  useNewUrlParser: true,
+  useFindAndModify: false,
+  useCreateIndex: true,
+  useUnifiedTopology: true
+};
+
+mongoose
+  .connect(url, options)
+  .then(() => {
+    console.log('MongoDB is connected');
+  })
+  .catch((err) => {
+    throw new Error(err);
+  });
 
 const LostSchema = new Schema({
   name: 'string',
